feat(jest): 100% coverage and branches for Logger.ts
